What are the sizes of all these files

+++ 0x00-hello_world/6-size.c
@@ -8,10 +8,10 @@
 
 int main(void)
 {
-	printf("Size of a char: %lu", sizeof(char));
-	printf("Size of an int: %lu", sizeof(int));
-	printf("Size of a long int: %lu", sizeof(long int));
-	printf("Size of a long long: %lu", sizeof(long long));
-	printf("Size of a long float: %lu", sizeof(long float));
+	printf("Size of a char: %d", (int) sizeof(char));
+	printf("Size of an int: %d", (int) sizeof(int));
+	printf("Size of a long int: %d", (int) sizeof(long int));
+	printf("Size of a long long: %d", (int) sizeof(long long));
+	printf("Size of a long float: %d",(int) sizeof(long float));
 	return (0);
 }
